MNIST Handwritten Digits/requirements.txt
streamlit
tensorflow
pillow
numpy
scikit-learn
matplotlib
seaborn
